Fixed a bug in double-responding because of hash change.
When you change slides, it starts animation and all, but also changes the hash, which the listener then listened to, and started
the animation and all again. We ended up with a duplicate of the incoming slide stacked on top of the outgoing slide.

So now we just check when the hash changes, if it's already the same as the current id, we don't do anything.