Don't think I can make classes into real Javascript constructors.
It is possibly to change _newObject to return a constructor which will correctly initialize it's objects when they are created.
However, a javascript constructor cannot (AFAIK) be used to create new instances of constructors. Then again, the Function constructor
exists, so I don't know. anyway, with out that, I can't do, e.g., `new Class()` to create a new constructor, only a new object.